net: sched: act_ct_4_18: Fix offload_timeout value to uint
It's being stored as uint so don't allow uint64 as big value could trim
to 0.

Issue: 2346931

+++ net/sched/act_ct.c
@@ -33,8 +33,8 @@
 #include <net/netfilter/ipv6/nf_defrag_ipv6.h>
 #include <uapi/linux/netfilter/nf_nat.h>
 
-static unsigned long offload_timeout = 30;
-module_param(offload_timeout, ulong, 0644);
+static uint offload_timeout = 30;
+module_param(offload_timeout, uint, 0644);
 MODULE_PARM_DESC(offload_timeout, "Flow offload timeout in seconds");
 
 static struct workqueue_struct *act_ct_wq;
